+++ apps/cookbook/tsconfig.json
@@ -53,7 +53,8 @@
       "@kirbydesign/designsystem/kirby-app": ["libs/designsystem/kirby-app/index.ts"],
       "@kirbydesign/designsystem/chart": ["libs/designsystem/chart/index.ts"],
       "@kirbydesign/designsystem/fab-sheet": ["libs/designsystem/fab-sheet/index.ts"],
-      "@kirbydesign/designsystem/data-table": ["libs/designsystem/data-table/index.ts"]
+      "@kirbydesign/designsystem/data-table": ["libs/designsystem/data-table/index.ts"],
+      "@kirbydesign/designsystem/reorder-list": ["libs/designsystem/reorder-list/index.ts"]
     },
     "target": "es2020"
   },
