+++ apps/flows/tsconfig.json
@@ -71,7 +71,8 @@
       "@kirbydesign/designsystem/chart": ["libs/designsystem/chart/index.ts"],
       "@kirbydesign/designsystem/helpers/scss": ["libs/designsystem/helpers/scss/index.ts"],
       "@kirbydesign/designsystem/fab-sheet": ["libs/designsystem/fab-sheet/index.ts"],
-      "@kirbydesign/designsystem/data-table": ["libs/designsystem/data-table/index.ts"]
+      "@kirbydesign/designsystem/data-table": ["libs/designsystem/data-table/index.ts"],
+      "@kirbydesign/designsystem/reorder-list": ["libs/designsystem/reorder-list/index.ts"]
     },
     "target": "es2020"
   },
