+++ libs/designsystem/reorder-list/ng-package.json
@@ -1,0 +1,1 @@
+{}
