+++ libs/designsystem/src/lib/components/index.ts
@@ -3,8 +3,6 @@ export { KirbyBadge as BadgeComponent } from './web-component-proxies.component'
 export { GridComponent } from './grid/grid.component';
 export { GridCardConfiguration } from './grid/grid-card-configuration';
 
-export * from './reorder-list';
-
 /* prettier-ignore */
 export { SegmentedControlComponent, SegmentedControlMode} from './segmented-control/segmented-control.component'; // See issue #1555 for prettier-ignore explanation
 export { SegmentItem } from './segmented-control/segment-item';
