+++ libs/designsystem/tsconfig.json
@@ -44,7 +44,8 @@
       "@kirbydesign/designsystem/kirby-app": ["kirby-app/index.ts"],
       "@kirbydesign/designsystem/chart": ["chart/index.ts"],
       "@kirbydesign/designsystem/fab-sheet": ["fab-sheet/index.ts"],
-      "@kirbydesign/designsystem/data-table": ["data-table/index.ts"]
+      "@kirbydesign/designsystem/data-table": ["data-table/index.ts"],
+      "@kirbydesign/designsystem/reorder-list": ["reorder-list/index.ts"]
     },
     "esModuleInterop": true,
     "target": "es2020"
