+++ libs/designsystem/tsconfig.spec.json
@@ -48,7 +48,8 @@
       "@kirbydesign/designsystem/chart": ["chart/index.ts"],
       "@kirbydesign/designsystem/helpers/scss": ["helpers/scss/index.ts"],
       "@kirbydesign/designsystem/fab-sheet": ["fab-sheet/index.ts"],
-      "@kirbydesign/designsystem/data-table": ["data-table/index.ts"]
+      "@kirbydesign/designsystem/data-table": ["data-table/index.ts"],
+      "@kirbydesign/designsystem/reorder-list": ["reorder-list/index.ts"]
     }
   },
   "files": ["test.ts"],
